♻️ REFACTOR: Move implementation specific extras code
`BackendEntityExtrasMixin` leaked implementation specific variables and functions.
This has now been moved to the `DjangoNode`/`SqlaNode` classes.
